requirements.txt
selenium
pytest
webdriver-manager
black
allure-pytest